+++ src/effects/specialFeats.js
@@ -28,6 +28,7 @@ import { vigilantBlessingEffect } from "./feats/vigilantBlessing.js";
 import { steadyAimEffect } from "./feats/steadyAim.js";
 import { shiftEffect } from "./feats/shift.js";
 import { blessedStrikesEffect } from "./feats/blessedStrikes.js";
+import { planarWarriorEffect } from "./feats/planarWarrior.js";
 
 export function baseFeatEffect(document, label) {
   return {
@@ -178,6 +179,10 @@ export async function featureEffectAdjustment(ddb, character, document) {
     }
     case "Indomitable": {
       document = indomitableEffect(document);
+      break;
+    }
+    case "Planar Warrior": {
+      document = await planarWarriorEffect(document);
       break;
     }
     case "Potent Cantrip": {